move modal into subfolder, typecheck script
