+++ client/src/components/Modal/Modal.tsx
@@ -1,7 +1,7 @@
 import useModal from "@/lib/hooks/useModal";
 import type { PropsWithChildren } from "react";
 import { useRef } from "react";
-import * as S from "./Modal.style";
+import S from "./Modal.style";
 
 type ModalProps = {
 	modalId: string;
